+++ CHANGELOG.md
@@ -6,6 +6,7 @@ GAMSPy 1.4.1
 - General
   - Fix implicit parameter validation bug.
   - Allow the usage of Container as a context manager.
+  - Migrate GAMSPy CLI to Typer.
 - Testing
   - Lower the number of dices in the interrupt test and put a time limit to the solve.
 - Documentation